+++ integrations/room/src/jvmTest/kotlin/com/powersync/integrations/room/Utils.jvm.kt
@@ -3,4 +3,5 @@ package com.powersync.integrations.room
 import androidx.room.Room
 import androidx.room.RoomDatabase
 
-actual fun createDatabaseBuilder(): RoomDatabase.Builder<TestDatabase> = Room.inMemoryDatabaseBuilder<TestDatabase>()
+actual fun createDatabaseBuilder(): RoomDatabase.Builder<TestDatabase> =
+    Room.inMemoryDatabaseBuilder<TestDatabase>().addCallback(TestDatabase)